make it compile (fix size of stat.st_nlink)
I don't really understand what's going on, but on my OS X machine, trying to install this package would result in an error message saying:

    VerificationError: function ev_stat_init: field 'stat.st_nlink' is declared as 8 bytes, but is really 2 bytes

So I just fixed this field to make it 2 bytes instead of 8. After making the change, everything went smoothly!

+++ pypycore.py
@@ -98,7 +98,7 @@ struct ev_child {
 };
 
 struct stat {
-    long long st_nlink;
+    short st_nlink;
     ...;
 };
